Fix win/lose: Add autoreveal and disable board clicking upon win/loss
Also fixed a bug where winning does not set mines counter to 0.
Reveals all the unrevealed tiles upon win/loss. Tells the user that the board
can no longer be interacted with and that they should start a new game. Added
row and column fields to tile to allow chording and future functions that
depend on this information. Thus, tiles can no longer be generated using map
function but instead be generated imperatively (using for loops)