Add timer and function that reveals surrounding zeroes and update chord
This process uses DFS to search for surrounding zeroes. This includes the case
when chording reveals a zero tile, so the function is updated to support this
functionality. The timer (that was previously used as a placeholder) now
correctly runs when the game starts and stops when the game ends.