Fix bug where revealZeroesAroundTile can be performed after win/loss.
Added the condition for that function. Also cleaned up a code a bit by
adding indentation and sticking to a convention for conditional checks:
only return immediately for meeting win/lose condition if the body block
is too long; else nest the body block inside if win/lose condition not
met.