Add short circuit logic to wasted actions and track useful/wasted clicks
A useful click is an action that reveals a tile or flags a mine that has
not been flagged before. Conversely, a wasted click is an action that
does not reveal any tile or flags a number/a tile that has been flagged
before. Added a few field in object called isFlaggedBefore. At the same
time, added short circuit logic to such wasted actions by returning the
function immediately, as we have to check for it anyways. This optimises
the code slightly.