Fix bug where find3BV runs into error after switching to a larger board
After switching to a larger board, height and width are updated but the
both setBoard and find3BV are within the same function resetBoard. Thus,
find3BV is using the new, larger height and width, while the board is
still the old, smaller board (note that setStates are async), leading to
out of bounds error. This can also happen in useState when find3BV runs
after height and width is updated. Thus, initialisation functions are
now used to initialise states. A new useEffect is added where it runs
when board is updated and before the game starts.