Add ability to show/hide metrics in Settings (saved in localStorage)
Added a form to allow such a feature, with the data being stored in
localStorage and retrieved upon page refresh. The state is lifted from
Settings to Navbar to App using callback functions, and is passed down
from App to GameBoard to GameResult via props. GameResult then uses the
data to conditionally render the metrics.