Style footer, more gap btw board & result, unmount navbar item correctly
Footer now is at the bottom of the screen with padding and background
color. Currently, it only contains the link to the repository, but more
features will be added to the footer. Added margin-bottom to board
container as gameResult was too close to the board. When no navbar item
is displayed, there is still a black line and an empty space at the top;
added dynamic styling and conditional rendering to resolve it.