Add results of all games to firebase and display them in AllGamesPlayed
Added firebase setup code and updated gameResult to add the raw data of
a game result to a collection, resultsCollection. The new component,
AllGamesPlayed, finds the computed data (e.g. 3BV/s) and displays both
the raw and computed data. These data is then rendered on the screen,
under Navbar. The advanced metrics, similar to gameResult, can be
shown/hidden under Settings. This involves passing showMetricsData from
App to Navbar to AllGamesPlayed. Added some styling for the rendered
items to be more readable, but more styles would be added in the future.
Updated Navbar to have general event for all buttons, instead of
separate events for each button. This is done by supplying a different
argument for the event of each button.