Prevent first move from resulting in a loss
Added a new state, finished1stClick, to denote if it is the first move.
It remains false upon reset until the gameStatus becomes onGoing.
If 1st move flags a tile, nothing is changed. If 1st move reveals a
mines, keep it unrevealed (to prevent dark red tile from appearing),
then note down the position in a global state, firstRevealedPos (is null
upon reset). Then a new useEffect, where if gameStatus becomes onGoing,
sets finished1stClick to true. Furthermore, if firstRevealedPos has
something, reset the board to a new board that does not put mines at
that revealed position (stored in firstRevealedPos), and set
usefulClicks to 0 (as 1 is used for revealing the mine, which should not
count). Then another new useEffect, where if finished1stClick becomes
true, if firstRevealedPos has something, call revealZeroesAroundTile or
revealTile depending on the value of the new tile (with same position as
that revealed position). Also, set the time to 0.001s if there is only 1
click performed.

Note: this implementation randomly shuffles the mines among the
remaining positions, unlike the usual version where the mine is shifted
to the topmost number, by row, then column.

In short, if first move reveals a mine, reset the board (with mines not
allowed to generate on that same position as the revealed mine), and
treat it as if the user revealed a tile in the same position in the new
board.