Remove isAutoRevealed as we can check for gameStatus in Tile instead
gameStatus is passed in as props, from GameBoard to Board to Tile. The
handling of logic is now handled in Tile; board is no longer updated
upon game win/loss. This reduces a field in the objects of board, but
more importantly, this more closely resembles the logic of rendering the
board of an old match under AllGamesPlayed, which would be added soon.
Also removed some redundant/commented out code that is no longer used.