Save boardData to firebase and update resultsCollection
This boardData allows the user to view the board status when the game
ended, from AllGamesPlayed. This provides a more comprehensive history
of the games the user played. As firebase does not allow nested array,
the workaround used is similar to that of generateMines: map the 2D
array to a 1D array, while storing the width to uniquely determine the
shape of the 2D array. To retrieve the data, this 1D array would then be
mapped back to 2D array using width. In doing so, the old data in
firebase is breaking the app, so resultsCollection is updated to a new
collection for the new data.