Update GameBoard and GameResult to show old game results upon btn click
Refactor AllGamesPlayed to have a new component, OldGameResult. Results
are now displayed row by row in a table (rather than using column by
column using flexboxes to centrally align the text). This is to allow
the view button to easily retrieve the data of its old game result. Also
shortened the results to a new format as the results in the old format
are no longer able to fit onto the screen in one line. The data is then
passed from OldGameResult to AllGamesPlayed to Navbar to App to modify
boardProperties of App (so that GameBoard is in the correct size to
display the old board), while some other data are passed down to
GameBoard for Header and GameResult to display the old results
correctly. Tile is updated to autoreveal tiles when viewing old game
results, and GameResult is updated to not add the old game result to
firestore.