Add dropdown to change tile size/number size in Settings
Created arrays inside Settings that are mapped to elements containing
the values. The elements would change tileSize and numberSize upon
click. tileSize and numberSize are passed up to Navbar, then App. App
sets its new state tileSize to the new value, which is then passed down
to GameBoard, then Board (which changes the width of the board
container), then Tile (changing the tile size and the number size).
Dropdown overlays other elements and has styles: the selected element is
highlighted in blue, while the hovered element is highlighted in light
grey. Note the strange behaviour of the browser; details are stated in
the comments of styles in Board. Also added some tolerance for imprecise
floating point representation (e.g. when comparing with tileSize to
determine selected value).