Add ability to change font size in AllGamesPlayed to fit the long table
This new feature allows all the columns to fit on the screen neatly, by
reducing the font size. Also allows larger font size for fewer columns.
The styles used is similar to that of the dropdowns in Settings.
Similar to the dropdowns in Settings, a state variable, oldGamesFontSize
is added, and the font size of the table depends on oldGamesFontSize.
The creation of the button and dropdown is analogous to the dropdown in
Settings, just with different names and values. In fact, it is simpler
as the states are all managed within AllGamesPlayed: no prop drilling
needed. (Also an update to the previous commit: similarly, this also
saves the settings in localStorage, and retrieves the data upon reload.)