Add a way to filter results by game mode, stop text overflow in dropdown
The filtering is done by sending a query to firebase. If the filter is
equal to "All", remain as it is, otherwise filter based on the selected
filter, the state gameModeFilter. Updated the onSnapShot useEffect to
run whenever gameModeFilter is updated. Edited CSS as the text in the
dropdown for filtering by game mode is overflowing.