Add ability to filter old games by difficulty in AllGamesPlayed
This is similar to filtering by game mode. To repeat what is done, add a
label with a button that shows/hides a dropdown. This label shows a new
state, difficultyFilter. Clicking the button displays a dropdown list of
elements, which is mapped from an array of values. Meanwhile, other
dropdowns should be hidden, which is managed by a state showDropdown
(set to that dropdown name, and use conditional rendering to render it).
Clicking an item sets difficultyFilter to its value, and hides the
dropdown (set showDropdown to empty string). The item matching the
difficultyFilter is highlighted in light blue, while the hovered item
is highlighted in light grey. Finally, when difficultyFilter is updated,
the onSnapShot in the useEffect will run, making new query that filters
the results based on the difficultyFilter. The queries are chained to
allow for multiple filters.