Move computed fields up to AllGamesPlayed to allow sorting/filtering
Computation of such fields is now moved up to AllGamesPlayed (i.e. the
functions). The results array is now mapped to objects with raw and
computed data, then mapped to OldGameResult components, with the fields
of the objects being passed down as props. OldGameResult is now merely
in charge of rendering and presenting the data is a more palatable way
(e.g. rounding to 3dp, nearest %, displaying the date in DD MM YY
hh:mm:ss format). This change is to allow AllGamesPlayed to have access
to the computed fields and sort/filter the objects, before rendering
them as OldGameResult components.