Move dropdown elements into new reusable component, LabelAndDropdown
The dropdown is repeated a few times in AllGamesPlayed, so this new
component would reduce repetition and keep the AllGamesPlayed organised.
In particular, generateDropdownElements is moved into the new component,
while the const elements are removed and would be computed within
LabelAndDropdown, to maintain separation of concerns. The fields that
are changed between different dropdowns, as well as other parent
information that is needed, are passed in as props. Note that this
component is only for AllGamesPlayed: the logic of the Settings
dropdowns are slightly different. Both dropdowns can be shown at the
same time, unlike in AllGamesPlayed, where only 1 dropdown can be shown
at a time. Also, there is number comparison for the dropdown in charge
of number size, which cannot be applied to Strings here.