Add ability to filter by board properties in AllGamesPlayed
Note that this is not custom exclusive; other difficulties can also
appear if the properties match. New states boardFilter and
newBoardFilter are added to keep track of these filters, with the
latter to store the form data that can be used to update boardFilter.
As usual, the fields in boardFilter are then used to filter out objects
in oldGameResults that do not satisfy the conditions. Reused the styles
in changeBoard, and increased the right margin to better separate the
different types of fields.