Add ability to sort by est time, 3BV/s and efficiency in AllGamesPlayed
A custom sort function is added to sort based on the selected metric.
The sort function checks for sortBy, converts it to the corresponding
field name (as sortBy is the displayed name, different from the field
name in code), then sort by ascending/descending order based on the
field name. The array notation is used to access the field as the field
name is a variable stored as a string. Items are then appended to the
items in the sort by dropdown.