Fix bug: sort by Date/Time sorts by efficiency. Add more sort by options
Updated sortFunction to return immediately if we sort by date/time.
Also added more checks with the added sort by options. In the sort by
dropdown, the advanced performance metrics would also be hidden if they
are unchecked in Settings, by conditionally adding them to an array,
then concatenating this array to the other items. Also added the sort
directly in the line declaring filterByMines to fix the issue of
filterByWidth.filter(...) is not a function. Also tweaked some styles:
in Settings, the checkboxes are spaced apart slightly to better
differentiate the labels for each of them; minimum height is applied
for everything above the footer instead of the game board.