Add styles to win/lose notif, highlight to viewed result and NF field
Long commit body - tying up loose ends and finalising the App

Added some styles to the notification above the New Game button. This
has been long forgotten and is added to make winning or losing feel a
bit more special (yet not too distracting - as users may want to play
multiple games quickly).

Upon clicking the view button, the result corresponding to it is
highlighted in light blue. This allows the user to clearly see which
result they are viewing. This is done by capturing the id to the
retrieveGameData, which is then passed to App, to allow ChangeBoard and
GameBoard to remove the light blue highlight (when the board is reset).
(In the meantime, as key cannot be passed down as props, a new prop for
OldGameResult is added, id, which is identical to key). More
specifically, when the gameStatus is set to "" (ChangeBoard changes the
properties, which is then passed to App, which is then passed down to
GameBoard, triggering the useEffect and thus resetBoard, which sets the
gameStatus to ""). This is to prevent the highlight from being removed,
as viewing an old result with different board properties triggers the
useEffect and calls resetBoard. However, the gameStatus would then be
set to "viewOldMatch", thus preventing the removal of the highlight
(which is done by clearing the oldGameData in GameBoard and thus App -
it is still there for retrieval in AllGamesPlayed and Firestore).

In the meantime, clicking the view button leads to a warning that tile
elements do not have a unique key prop. This is due to Board taking the
id field of the board objects instead of key, so the Firestore board
objects now have field renamed to id (instead of key). This resulted in
a name change for OldGameResults in AllGamesPlayed. The results are now
outdated, so a new collection is used. The indexes are also re-added for
this new collection.

A new (computed) column in AllGamesPlayed is added: No Flags field. It
would be set to "Yes" if rightClicks (wasted or not) are used, otherwise
"No". A new LabelAndDropdown is created to filter for this field in
AllGamesPlayed.

Other bugs were fixed: in OldGameResult, the line
const height = props.width is corrected; wasted clicks is now sorted in
ascending rather than descending order; if the time is 12am, the hours
are now displayed in 00: rather than 0:; localStorage for oldGameData
is removed as it is redundant. Also slightly optimised the filter to
check the filter state once, instead of for each result.